Refactor swapping algorithm
